Now writes blueprints with version 23.
I made an assumption that the data after the zipped block, that is now identified with a length and a set of flags, is followed by a second block that could also be zipped. So far I have only seen unzipped data here though. I have not tried to save this section zipped, but I don't think that my understanding of this section is good enough to test yet.

The reader does not treat it this way. It still parses "matricesLength" and "matrixUnknown01" instead of what I think they are: "Data length" and "Data type flags" respectively.

I still don't understand any of the data stored in this huge new section.